UI: Change how color is stored with some fucntional changes in rendering
This contains lot of little refactor starting to store `Color` in simple u16 and using
a matrix for storing pixels instead of plan vector. To me seems more strainghtforward for
future implementation.
The result right now is same as before but more changes will happen here in the next future :).

Signed-off-by: Federico Guerinoni <guerinoni.federico@gmail.com>