UI: Use a fixed vector to keep ASM generated
In this way UI is very fluid, before was very laggy at some point,
because of the string that need to be render for all ASM generated since
time 0. Now I created a little data struct that answer to our need in
order to keep only the last 1k instructions. Seems a reasonable number.
For future improvement we can enable also a possibility to save as file.

Signed-off-by: Federico Guerinoni <guerinoni.federico@gmail.com>