Avoid having `Arc<Mutex<Bus>>` inside `Arm7tdmi`
The main idea of this change is to avoid having `Arc<Mutex<Bus>>`
inside `Arm7tdmi` and have just a `Bus` instead.
Having an ArcMutex it was causing me some issues while
working on interrupts implementation. Plus it's always been something
I was not 100% sure about. I changed the `Gba` logic
to hold only a reference to the cpu instead of both cpu
and bus.