Memory: Move `LCD` out of `InternalMemory`
The idea is to have hardware components directly attached
to the `Bus`. In the long term we should move all other components
to the `Bus`.
The solution we had before it's not suitable because an hardware component
doesn't have just registers but it also comes with
its own logic obiously.